UPDATE: Makefile

+++ Libft/ft_memset.c
@@ -6,21 +6,21 @@
 /*   By: kkurita <kkurita@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/18 17:57:39 by kkurita           #+#    #+#             */
-/*   Updated: 2021/01/18 17:57:39 by kkurita          ###   ########.fr       */
+/*   Updated: 2021/01/18 23:02:21 by kkurita          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include"libft.h"
+#include "libft.h"
 
-void *ft_memset(void *addr, int ch, size_t n)
+void	*ft_memset(void *addr, int ch, size_t n)
 {
-    int i;
+	size_t i;
 
-    i = 0;
-    while(i < n)
-    {
-        *((int *)addr + i) = ch;
-        i++;
-    }
-    return (addr);
+	i = 0;
+	while (i < n)
+	{
+		*((int *)addr + i) = ch;
+		i++;
+	}
+	return (addr);
 }
